Removed annoying debug print statements, added new bullet sprite

+++ source/core/src/main/com/csse3200/game/components/CombatStatsComponent.java
@@ -16,7 +16,7 @@ public class CombatStatsComponent extends Component {
   private static final Logger logger = LoggerFactory.getLogger(CombatStatsComponent.class);
   private int health;
   private int baseAttack;
-  private int attackTimer;
+  private float attackTimer;
 
   public CombatStatsComponent(int health, int baseAttack) {
     setHealth(health);
@@ -63,7 +63,7 @@ public class CombatStatsComponent extends Component {
    *
    * @param attackTimer attackTimer
    */
-  public void setAttackTimer(int attackTimer) {
+  public void setAttackTimer(float attackTimer) {
     if (attackTimer >= 0) {
       this.attackTimer = attackTimer;
     } else {
@@ -105,7 +105,8 @@ public class CombatStatsComponent extends Component {
   }
 
   /**
-   * Allows the entity to be hit by some attacker and deal some damage
+   * Allows the entity to be hit by some attacker and deal some damage, if they have waited
+   * for the designated time between attacks.
    * @param attacker the entity attacking
    */
   public void hit(CombatStatsComponent attacker) {
@@ -118,4 +119,6 @@ public class CombatStatsComponent extends Component {
       }
     }
   }
+
+
 }
